Use more lightweight dependencies. (#6)
Refactored SeerApiBuilder to an internal Builder class in SeerApi.
